ci: stabilize Rust and integration gates

+++ src/services/event_bus/tests/redis.rs
@@ -159,7 +159,7 @@ async fn redis_monitor_feed_preserves_order_dedup_and_scope() {
     // The maximum accepted A&D flag batch in another game must not enter
     // this target/game receiver's bounded history.
     for cursor in 0..100 {
-        first_replica.publish(received_game_event(8, cursor as i32, cursor as i64));
+        first_replica.publish(received_game_event(8, cursor, cursor as i64));
     }
     for (id, cursor) in [(31, 101), (32, 102), (33, 103)] {
         first_replica.publish(received_game_event(7, id, cursor));
